requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
pydantic>=2.0.0
